fix(nika-cadence): named-beat tick policy lives with the pure grammar
R1 measured 222 LOC of file-truth + D6 + clock classification that
does not open a path. nika-cadence is L0; emit and the firer must
share one v0_unsupported set. nika-arm keeps lock, claim, pin, and
run. Teaching lines stay byte-identical.

Also pin the ubuntu public-api snapshots CI already rendered
(nika-check PartialFaces from #1062; nika-cli zvariant blanket).
